Support Section type for dash docsets in tests
I did this to support a pull request I worked up in jazzy that adds type
information for Overview items so that they are recognized at a Section
type in the Dash docset index generation process. That change is
necessary because Dash requires that the index have types for everything
in the index.

All of the docSet.dsidx.csv changes I am confident are correct and
should be there. The apple_ref anchors on the other hand is something I
am not super familiar with. The addition of them hasn't seemed to break
anything as far as I can tell. Also, based on my very brief
understanding of the apple_ref anchors it seems that they should have
been there all along. If that is not the case please enlighten me as to
how it is supposed to work and why, and I will dig in and try and get it
there.